Move spec runner into it's own class, add some checks around re-unactioning a note
